pycask updated

+++ home/dotfiles/pycask.py
@@ -11,21 +11,24 @@ import argparse
 import sqlite3
 import pandas as pd
 import time
+
 pd.options.mode.copy_on_write = True
 
-Cask = namedtuple("Cask", "token name desp version homepage installed update url sha256")
+Cask = namedtuple(
+    "Cask", "token name desp version homepage installed update url sha256"
+)
 version_code_map = {
-    '10.11': 'el_caption',
-    '10.12': 'sierra',
-    '10.13': 'high_sierra',
-    '10.14': 'mojove',
-    '10.15': 'catalina',
-    '11': 'big_sur',
-    '12': 'monterey',
-    '13': 'ventura',
-    '14': 'sonoma'
+    "10.11": "el_caption",
+    "10.12": "sierra",
+    "10.13": "high_sierra",
+    "10.14": "mojove",
+    "10.15": "catalina",
+    "11": "big_sur",
+    "12": "monterey",
+    "13": "ventura",
+    "14": "sonoma",
 }
-code_version_map = {v:k for k,v in version_code_map.items()}
+code_version_map = {v: k for k, v in version_code_map.items()}
 
 
 class PyCask:
@@ -41,10 +44,10 @@ class PyCask:
 
         self.df_all = None
         self.jsons = None
-        
+
         if not os.path.exists(self.config_path):
             os.makedirs(self.config_path)
-            
+
         if os.path.exists(self.json_file):
             stat = os.stat(self.json_file)
             if time.time() - stat.st_mtime < 3600:
@@ -52,53 +55,51 @@ class PyCask:
 
         if os.path.exists(self.db_fname):
             self.from_sql_db()
-            
+
         if self.jsons is None:
             self.get_refresh()
-            
+
         if self.df_all is None:
             self.build_df()
-        
 
     def to_sql_db(self, name="casks"):
         conn = sqlite3.connect(self.db_fname)
         self.df_all.to_sql(name, conn, index=True, if_exists="replace")
         conn.close()
-        
+
     def from_sql_db(self, name="casks"):
         conn = sqlite3.connect(self.db_fname)
         self.df_all = pd.read_sql(f"select * from {name}", conn)
         self.df_all = self.df_all.set_index("token")
         conn.close()
-        
-    
+
     def get_version_code(self):
         version = platform.mac_ver()[0]
-        code = 'sonoma'
+        code = "sonoma"
         for v in version_code_map:
             if version.startswith(v):
                 code = version_code_map[v]
                 break
         return code
-    
+
     def get_refresh(self):
         print("refresh database")
         res = requests.get(self.api)
         if res.status_code == 200:
             self.jsons = json.loads(res.content.decode())
-            f = open(self.json_file, 'w')
+            f = open(self.json_file, "w")
             f.write(res.content.decode())
             self.build_df()
         else:
             print("return error")
-    
+
     def load_json(self):
         self.jsons = json.load(open(self.json_file))
 
     def get_json_results(self):
         results = []
         for cask in self.jsons:
-            
+
             token = cask["token"]
             name = cask["name"][0]
             desc = cask["desc"]
@@ -106,85 +107,82 @@ class PyCask:
             url = cask["url"]
             sha256 = cask["sha256"]
             installed = 0
-            #remove = cask["zap"][
-            #]
+            # remove = cask["zap"][
+            # ]
             version = cask["version"]
             if self.arch != "arm64":
                 variations = cask["variations"]
                 url_code = variations.get(self.code, {}).get("url", "")
                 sha256_code = variations.get(self.code, {}).get("sha256", "")
 
-                url = (url_code if len(url_code) > 0 else url)
-                sha256 = (sha256_code if len(sha256_code) > 0 else sha256)
+                url = url_code if len(url_code) > 0 else url
+                sha256 = sha256_code if len(sha256_code) > 0 else sha256
             c = Cask(token, name, desc, version, homepage, installed, None, url, sha256)
             results.append(c)
         return results
-    
+
     def build_df(self):
-        if self.df_all is None:
-            update = False
-        else:
-            update = True
         results = self.get_json_results()
-        if not update: #initial 
+        if self.df_all is None:  # initial, no db, df_all not initialized
             self.df_all = pd.DataFrame(results)
             self.df_all = self.df_all.set_index("token")
-            #print(self.df_all)
+            # print(self.df_all)
             self.to_sql_db()
         else:
             for cask in results:
                 token = cask.token
                 if token not in self.df_all.index:
-                    #new_casks.append(cask)
-                    #print(token)
-                    #print(cask)
-                    self.df_all.loc[token] = cask[1:] 
-                    #self.to_sql_db()
+                    # new_casks.append(cask)
+                    # print(token)
+                    # print(cask)
+                    self.df_all.loc[token] = cask[1:]
+                    # self.to_sql_db()
                 else:
                     df_pkg = self.df_all.loc[token]
 
                     if df_pkg.version != cask.version:
-                        print(f"{cask.token}\r\t\t\t{df_pkg.version.split(',')[0]}\r\t\t\t\t---->\t{cask.version.split(',')[0]}")
-                        #df_pkg.url = cask.url
+                        print(
+                            f"{cask.token}\r\t\t\t{df_pkg.version.split(',')[0]}\r\t\t\t\t---->\t{cask.version.split(',')[0]}"
+                        )
                         self.df_all.loc[token, "url"] = cask.url
-                        #df_pkg.sha256 = cask.sha256
                         self.df_all.loc[token, "sha256"] = cask.sha256
+
                         if df_pkg.installed == 0:
-                            #df_pkg.version = cask.version
                             self.df_all.loc[token, "version"] = cask.version
                         else:
-                            #df_pkg.update = cask.version
                             self.df_all.loc[token, "update"] = cask.version
             self.to_sql_db()
-                
-                        
+
     def search(self, key):
-        #results = []
+        # results = []
         if len(key) < 3:
             print("search key must >= 3")
-            return []
+            return None
 
         df = self.df_all.query(f"token.str.contains('{key}')")
         return df
-            
 
     def print_cask(self, cask):
-        print(f"{cask.installed} {cask.name}\r\t\t\t{cask.version}\r\t\t\t\t\t{cask.desp}")
-        
+        print(
+            f"{cask.installed} {cask.name}\r\t\t\t{cask.version.split(',')[0]}\r\t\t\t\t\t{cask.desp}"
+        )
 
-    def download(self, c, token, upgrade=False):
+    def download(self, cask, token, upgrade=False):
+        if cask["installed"] == 1 and cask["update"] == cask["version"]:
+            print(f"{token} with version {cask['version']} alreay installed")
+            return True
         print(f"starting download {token}")
-        url = c.url
+        url = cask.url
         dest = tempfile.gettempdir()
         f = os.path.basename(url)
-        command = ["aria2c", "-c", url,  "-d", dest, "-o", f]
+        command = ["aria2c", "-c", url, "-d", dest, "-o", f]
         print(command)
         popen = subprocess.Popen(command, stdout=sys.stdout)
         out = popen.wait()
         if upgrade and out == 0:
             out == self.remove(token, upgrade)
-        if out == 0: #success
-            version_new = (c["update"] if upgrade else c["version"])
+        if out == 0:  # success
+            version_new = cask["update"] if upgrade else cask["version"]
             self.install(os.path.join(dest, f), token=token, version=version_new)
         else:
             print("download failed")
@@ -198,44 +196,45 @@ class PyCask:
             format = "dmg"
 
         getattr(self, f"install_{format}")(fname)
-        
+
         print(f"{token} installed with version {version}")
         self.df_all.loc[token, "installed"] = 1
         self.df_all.loc[token, "version"] = version
         self.to_sql_db()
-
 
     def remove(self, token, upgrade=False):
         # api = f"{self.api_base}/cask/{token}.json"
         # res = requests.get(api)
         # if res.status_code == 200:
         #     jsons = json.loads(res.content)
-        # 
+        #
         # else:
         #     pass
+        jsons = None
         for cask in self.jsons:
             if cask["token"] == token:
                 jsons = cask
                 break
+        if jsons is None:
+            print(f"App with name = {token} not found")
+            return
         artifacts = jsons["artifacts"]
         for artifact in artifacts:
             apps = artifact.get("app", [])
             for app in apps:
                 os.system(f"rm -rf ~/Applications/{app}")
-            if not upgrade: # when upgrade, not zap apps
+            if not upgrade:  # when upgrade, not zap apps
                 zaps = artifact.get("zap", [])
                 for zap in zaps:
                     trashes = zap.get("trash", [])
                     for trash in trashes:
                         os.system(f"rm -rf {trash}")
-        if not upgrade: # when upgrade, not update db when remove
+        if not upgrade:  # when upgrade, not update db when remove
             self.df_all.loc[token, "installed"] = 0
             self.df_all.loc[token, "update"] = None
             self.to_sql_db()
-                
-            
-        
-    #def install_dmg(self, fname):
+
+    # def install_dmg(self, fname):
     #    from dmginstall import installDmg
     #    installDmg(fname)
 
@@ -284,60 +283,89 @@ class PyCask:
             for fmt in formats:
                 cmd = fmt_path.format(depth, fmt)
                 output = self.get_output(cmd).strip()
-                if output != '':
-                    if depth == 1 and fmt == 'dmg':
+                if output != "":
+                    if depth == 1 and fmt == "dmg":
                         self.install_dmg(output)
                         continue
                     else:
-                        if fmt == 'app':
+                        if fmt == "app":
                             self.copy_app(output)
                             return
-                        elif fmt == 'pkg':
+                        elif fmt == "pkg":
                             os.system(f'sudo installer -pkg "{output}" -target /')
                             return
         print("No usable files found")
-        
+
     def main(self):
         if self.args.search:
             res = self.search(self.args.search)
+            if res is None:
+                return
             if len(res) == 0:
-                print(f"Not found {self.args.search}")
+                print(f"Not found App contains {self.args.search}")
             else:
                 for idx in res.index:
                     cask = res.loc[idx]
-                    #print(cask)
+                    # print(cask)
                     self.print_cask(cask)
 
-
-                    
         elif self.args.install:
 
             token = self.args.install
             res = self.search(token)
+            if res is None:
+                return
 
             if len(res) == 0:
                 print(f"Package with name ({token}) not found")
-                
+
             elif token in res.index:
                 cask = res.loc[token]
                 out = self.download(cask, token)
                 if out == 0:
                     print(f"Installing. {token} Success!")
             else:
-                print("There are multiple casks:")
+                
                 for idx in res.index:
                     self.print_cask(res.loc[idx])
- 
-        elif self.args.update:
+                num = input(f"There are multiple casks, install which one :(1--{len(res.index)})")
+                num = int(num)
+                if num > 0 and num < len(res.index):
+                    token = res.index[num]
+                    self.download(res.loc[idx], token)
+
+        elif self.args.upgrade:
             stat = os.stat(self.json_file)
-            if time.time() - stat.st_mtime > 600: # 10min
+            if time.time() - stat.st_mtime > 600:  # 10min
                 self.get_refresh()
-            update_pkgs = self.df_all.query("update.notnull() and update != version")
-            for idx in update_pkgs.index:
-                pkg = update_pkgs.loc[idx]
-                print(f"{idx}\r\t\t{pkg['name']}\r\t\t\t\t{pkg.version.split(',')[0]}\r\t\t\t\t\t\t--->\t{pkg['update'].split(',')[0]}")
-                self.download(pkg, idx, upgrade=True)
-                
+            if self.args.upgrade == "*": # upgrade all pkgs
+                update_pkgs = self.df_all.query("update.notnull() and update != version")
+                if len(update_pkgs) == 0:
+                    return
+                for idx in update_pkgs.index:
+                    pkg = update_pkgs.loc[idx]
+                    print(
+                        f"{idx}\r\t\t{pkg['name']}\r\t\t\t\t{pkg.version.split(',')[0]}\r\t\t\t\t\t\t--->\t{pkg['update'].split(',')[0]}"
+                    )
+                choice = input("upgrade all these apps? (y/n)")
+                if choice.lower().strip() == "y":
+                    for idx in update_pkgs.index:
+                        pkg = update_pkgs.loc[idx]
+                        self.download(pkg, idx, upgrade=True)
+            else:
+                token = self.args.upgrade
+                if token in self.df_all.index:
+                    pkg = self.df_all.loc[token]
+                    if pkg["installed"] == 0:
+                        print(f"{token} not installed, cannot upgrade")
+                    elif pkg["update"] is None or pkg["update"] == pkg["version"]:
+                        print(f"{token} No need to upgrade")
+                    elif pkg["update"] != pkg["version"]:
+                        self.download(pkg, token, upgrade=True)
+                        
+                else:
+                    print(f"No app named {token}")
+
         elif self.args.list:
             self.list_pkgs()
 
@@ -351,25 +379,26 @@ class PyCask:
                 self.to_sql_db()
             else:
                 print(f"Package with {token} not found")
-            
+
+
 if __name__ == "__main__":
     import os
     import sys
+
     parser = argparse.ArgumentParser("search cask to install")
-    
+
     parser.add_argument("-s", "--search", type=str, help="search apps")
     parser.add_argument("-l", "--list", action="store_true", help="list apps")
     parser.add_argument("-i", "--install", type=str, help="install apps")
     parser.add_argument("-r", "--remove", type=str, help="remove apps")
-    parser.add_argument("-u", "--update", action="store_true", help="update database")
+    parser.add_argument("-u", "--upgrade", type=str, nargs="?", const="*", help="update apps or for all")
     parser.add_argument("-m", "--marked", type=str, help="mark a package as installed")
 
     args = parser.parse_args()
-
     cask = PyCask(args)
     cask.main()
-    #if len(sys.argv) >= 2:
+    # if len(sys.argv) >= 2:
     #    res = cask.search(sys.argv[1])
     #    #cask.download(res[0])
-    #else:
+    # else:
     #    print("you should run with ./pycask.py xxxx")
